修复重复初始化json导致内存泄露。 修复cel_reallocate中old_size未初始化导致内存泄露。

+++ cel/src/net/http.c
@@ -1012,7 +1012,7 @@ int cel_httpbodycache_read(CelHttpBodyCache *cache,
 		memcpy(buf, cel_stream_get_buffer(&(cache->buf)) + first, _size);
 		size = _size;
 	}
-	if (size < buf_size) 
+	if (size < (int)buf_size) 
 		*((char*)buf + size) = '\0';
 
     return size;
